+++ otc/src/tests/mod.rs
@@ -17,4 +17,5 @@ pub mod mock;
 
 pub mod cancel_order;
 pub mod fill_order;
+pub mod partial_fill_order;
 pub mod place_order;
